refactor(ee): extract Stripe subscription cancellation into one helper
Cancelling at Stripe hinges on a subtlety worth stating once: "already canceled"
comes back as a 400 with no distinguishing code, and reading it as a failure turns
an idempotent retry into a permanent one. That classification lived privately in
org-cancellation.ts; the duplicate-checkout reconciliation needs the same call, and
a second copy of a Stripe error-message regex is the kind of debt that rots apart.

`stripe/cancel.ts` now owns it, taking the caller's reason as log context instead of
baking "for a deleted org" into the message.
